Move `shouldSkipFile` to `path_utils.dart`
